@@ -44,25 +44,24 @@
         "luxon": "^1.11.3"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~0.6.5",
-        "@angular/animations": "^7.0.3",
-        "@angular/cli": "^7.0.5",
-        "@angular/common": "^7.0.3",
-        "@angular/compiler": "^7.0.3",
-        "@angular/compiler-cli": "^7.2.7",
-        "@angular/core": "^7.0.3",
-        "@angular/forms": "^7.0.3",
-        "@angular/http": "^7.0.3",
-        "@angular/language-service": "^7.0.3",
-        "@angular/platform-browser": "^7.0.3",
-        "@angular/platform-browser-dynamic": "^7.0.3",
-        "@angular/router": "^7.0.3",
+        "@angular-devkit/build-angular": "~0.800.0",
+        "@angular/animations": "^8.0.0",
+        "@angular/cli": "^8.0.0",
+        "@angular/common": "^8.0.0",
+        "@angular/compiler": "^8.0.0",
+        "@angular/compiler-cli": "^8.0.0",
+        "@angular/core": "^8.0.0",
+        "@angular/forms": "^8.0.0",
+        "@angular/language-service": "^8.0.0",
+        "@angular/platform-browser": "^8.0.0",
+        "@angular/platform-browser-dynamic": "^8.0.0",
+        "@angular/router": "^8.0.0",
         "@types/jasmine": "~2.8.3",
         "@types/jasminewd2": "~2.0.2",
         "@types/luxon": "^1.11.1",
         "@types/node": "~6.0.60",
         "codecov": "^3.1.0",
-        "codelyzer": "^4.0.1",
+        "codelyzer": "^5.0.1",
         "core-js": "^2.4.1",
         "jasmine-core": "~2.8.0",
         "jasmine-spec-reporter": "~4.2.1",
@@ -72,17 +71,17 @@
         "karma-jasmine": "~1.1.0",
         "karma-jasmine-html-reporter": "^0.2.2",
         "luxon": "^1.11.3",
-        "ng-packagr": "^4.7.1",
-        "ngx-material-timepicker": "^3.0.2",
-        "node-sass": "^4.11.0",
+        "ng-packagr": "^5.2.0",
+        "ngx-material-timepicker": "^3.2.2",
+        "node-sass": "^4.12.0",
         "npm-run-all": "^4.1.3",
         "protractor": "^5.4.2",
-        "rxjs": "^6.3.3",
+        "rxjs": "^6.5.2",
         "travis-deploy-once": "^5.0.0",
         "ts-node": "~4.1.0",
-        "tsickle": "^0.34.0",
+        "tsickle": "^0.35.0",
         "tslint": "~5.9.1",
-        "typescript": "3.1.6",
-        "zone.js": "^0.8.26"
+        "typescript": "3.4.5",
+        "zone.js": "~0.9.1"
     }
 }
