@@ -285,4 +285,4 @@ export const spyOnFunction = <T>(obj: T, func: keyof T) => {
     const spy = jasmine.createSpy(func as string);
     spyOnProperty(obj, func, 'get').and.returnValue(spy);
     return spy;
-  };
+};
